Extracts and stores event data for AEP
Adds a script to extract event data from the `window._eventData.apiResponse` object, which is populated by the fetch event data rule action.

Also introduces a constant for the 'eventDataReady' event name to ensure consistency between the script that fetches the data and the script that extracts it. This promotes modularity and reduces the risk of typos.